requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2        
catboost==1.2.7
category_encoders==2.7.0 
certifi==2025.1.31       
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1       
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dash==3.0.2
debugpy==1.8.13
decorator==5.2.1
deprecation==2.1.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.38.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.1.4
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.7
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt_toolkit==3.0.50
protobuf==6.30.2
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycaret==3.3.2
pydeck==0.9.1
Pygments==2.19.1
pyod==2.0.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smmap==5.0.2
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.45.0
stumpy==1.13.0
sympy==1.14.0
tbats==1.1.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
tsfresh==0.21.0
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.130
ultralytics-thop==2.0.14
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xxhash==3.5.0
yellowbrick==1.5
zipp==3.21.0